requirements.txt
google-adk>=0.4.0
google-cloud-aiplatform>=1.38.0
python-dotenv